Refactor #31: change TS compilation target to ES6 for entire project
Using a compilation target other than ES3 or ES5 allows for extending
the Error class without having to manually set the prototype for the
custom errors. We will introduce such custom Errors in the next commits.

Moreover, ES6 is supported by all modern browsers, so there is no reason
not to use it.